pyproject.toml
[project]
name = "call-me-maybe"
version = "0.15.3"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.12.5",
    "python-magic>=0.4.27",
    "llm_sdk",
]

[dependency-groups]
dev = [
    "flake8>=7.3.0",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
]
[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
python_files = ["test*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.uv.workspace]
members = ["llm_sdk"]

[tool.uv.sources]
llm_sdk = { workspace = true }